scripts/requirements.txt
python-dateutil==2.8.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
